FIX some warning (external dashboard)

+++ Core/Algorithms/mitkDataTreeNodeFactory.cpp
@@ -745,8 +745,8 @@ void mitk::DataTreeNodeFactory::SetDefaultImageProperties(mitk::DataTreeNode::Po
   mitk::TransferFunction::Pointer tf = mitk::TransferFunction::New();
   if(image.IsNotNull())
   {
-    int m_Min = image->GetScalarValueMin();
-    int m_Max = image->GetScalarValueMax();
+    float m_Min = image->GetScalarValueMin();
+    float m_Max = image->GetScalarValueMax();
     tf->GetScalarOpacityFunction()->Initialize();
     tf->GetScalarOpacityFunction()->AddPoint ( m_Min, 0 );
     tf->GetScalarOpacityFunction()->AddPoint ( m_Max, 1 );
